Housing Data
Worked as an Analyst for an investment bank. The team I was working with wanted to understand how it should allocate dollars earmarked for investment into mortgage-backed securities. They've asked me to look into the factors that drive home prices. 

The project was worked on Microsoft Excel, and I looked into 4 factors I thought would drive the price of a house.